Add built in smolagents Gradio app
