+++ track_explorer/analysts/smolagents_analyst/smolagents_analyst.py
@@ -3,7 +3,7 @@
 import logging
 from typing import Any, Callable, Dict, Generator, List, Optional
 
-from smolagents import ChatMessage, CodeAgent, stream_to_gradio
+from smolagents import ChatMessage, CodeAgent, GradioUI, stream_to_gradio
 from sqlalchemy import Engine, create_engine, inspect
 
 from ..base_analyst import BaseAnalyst
@@ -100,3 +100,15 @@ class SmolAgentsAnalyst(BaseAnalyst):
             new_messages.append(msg)
             yield new_messages
         yield new_messages
+
+    def launch(self) -> None:
+        """Launches the analyst."""
+        logger.info("launch | Launching SmolAgentsAnalyst")
+        if not self.initialized:
+            self._initialize()
+            self.initialized = True
+        gradio_ui = GradioUI(
+            agent=self.agent,
+        )
+        gradio_ui.launch()
+        logger.info("launch | SUCCESS")
